Split off `Foreign.JavaScript.CallJavaScript`
